Issue-04 - tested and working when headphones are removed
