+++ app/src/main/java/com/example/mike/mp3player/service/AudioFocusManager.java
@@ -1,22 +1,17 @@
 package com.example.mike.mp3player.service;
 
-import android.content.BroadcastReceiver;
 import android.content.Context;
-import android.content.Intent;
-import android.content.IntentFilter;
 import android.media.AudioManager;
 
 import androidx.media.AudioAttributesCompat;
 import androidx.media.AudioFocusRequestCompat;
 import androidx.media.AudioManagerCompat;
 
-public class AudioFocusManager extends BroadcastReceiver
+public class AudioFocusManager
  implements AudioManager.OnAudioFocusChangeListener {
 
     public static final float MEDIA_VOLUME_DEFAULT = 1.0f;
     private static final float MEDIA_VOLUME_DUCK = 0.2f;
-    private static final IntentFilter AUDIO_NOISY_INTENT_FILTER =
-            new IntentFilter(AudioManager.ACTION_AUDIO_BECOMING_NOISY);
 
     MyMediaPlayerAdapter player;
     AudioManager audioManager;
@@ -98,28 +93,5 @@ public class AudioFocusManager extends BroadcastReceiver
                 .build();
         int result = AudioManagerCompat.abandonAudioFocusRequest(audioManager, audioFocusRequestCompat);
         return  isRequestGranted(result);
-    }
-
-    @Override
-    public void onReceive(Context context, Intent intent) {
-        if (AudioManager.ACTION_AUDIO_BECOMING_NOISY.equals(intent.getAction())) {
-            if (player.isPlaying()) {
-                player.pause();
-            }
-        }
-    }
-
-    public void registerAudioNoisyReceiver() {
-        if (!audioNoisyReceiverRegistered) {
-            context.registerReceiver(this, AUDIO_NOISY_INTENT_FILTER);
-            audioNoisyReceiverRegistered = true;
-        }
-    }
-
-    public void unregisterAudioNoisyReceiver() {
-        if (audioNoisyReceiverRegistered) {
-            context.unregisterReceiver(this);
-            audioNoisyReceiverRegistered = false;
-        }
     }
 }
